requirements.txt
# Gmail
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# HTTP and HTML
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1

# Optional: LLM for summarization and category inference
# anthropic>=0.18.0
# openai>=1.0.0
